requirements.txt
python-dotenv==1.1.1
Jinja2==3.1.6
livereload==2.7.1 
more-itertools==10.7.0